[graphics/jasper2] Disable parallel build when building docs.
It is possible for `doxygen` to get into a race where its output
directory doesn't yet exist, so it attempts to create it, but in between
these two actions, another process creates it, so `mkdir(2)` fails with
`EEXIST` and `doxygen` bails. This should probably be fixed in
`doxygen`; until then, don't use parallel builds with it.