[devel/cmake] Better architecture check for Apple platforms.
Recent CMake config files contain a check to ensure that the package
supports the same architecture as its user. Currently, it checks only
the pointer bitness.

On Darwin, we can do better. We know the exact architectures that the
package and its user were built for. We can check, then, that the user's
architectures are a subset of those of the package.