[science/kicad] Fix the broken RTTI.
KiCad currently links everything into each KiFace (plugin), which
results in multiple definitions of RTTI data. On macOS 11 and up, this
isn't a problem, because `libc++abi` there is built with
`_LIBCXXABI_FORGIVING_DYNAMIC_CAST`. On older versions, though, this
causes weird failures because `dynamic_cast`s fail due to the multiple
RTTI definitions. Note that exceptions are actually broken as well even
on macOS 11 and up, because `_LIBCXXABI_FORGIVING_DYNAMIC_CAST` only
applies to `dynamic_cast`, not exception handler matching.

The _right_ solution is to link RTTI once, in the main binaries and in
the Python plugins; the KiFaces will then use the RTTI data from them. A
more pragmatic solution for now is to give the RTTI data default
visibility and weak linkage, so that `dyld(1)` will coalesce the
references at runtime. I've implemented this second solution, but MP
upstream really wants the first one implemented in KiCad upstream.